+++ R/add_ref.R
@@ -13,7 +13,7 @@ add_title <- function(vis, ...) {
 #' Add a config section to a vega plot
 #'
 #' @param vis a ggvis object
-#' @param ... vega title properties
+#' @param ... vega config properties
 #' @return a ggvis object
 #' @export
 add_config <- function(vis, ...) {
@@ -24,18 +24,18 @@ add_config <- function(vis, ...) {
 #' Add a mark
 #'
 #' @param vis a ggvis object
-#' @param ... vega title properties
+#' @param ... vega mark properties
 #' @return a ggvis object
 #' @export
 add_mark_ <- function(vis, ...) {
   mark <- vega_mark(...)
 
   #
-  if (!"mark" %in% names(vis$vega)) vis$vega$mark <- list()
+  if (!"marks" %in% names(vis$vega)) vis$vega$marks <- list()
 
   # check if named marks exist in both the incoming mark and vis
   if ("name" %in% names(mark)) {
-    locs <- which(mark$name %in% unlist(purrr::map(vis$vega$mark, "name")))
+    locs <- which(mark$name %in% unlist(purrr::map(vis$vega$marks, "name")))
   } else {
     locs <- c()
   }
@@ -43,10 +43,89 @@ add_mark_ <- function(vis, ...) {
   # if there's any overlap in names, replace the mark, otherwise add the mark
   lapply
   if (length(locs) > 0) {
-    vis$vega$mark[locs] <- list(mark)
+    vis$vega$marks[locs] <- list(mark)
   } else {
-    vis$vega$mark[[length(vis$vega$mark) + 1]] <- mark
+    vis$vega$marks[[length(vis$vega$mark) + 1]] <- mark
   }
+
+  # return
+  vis
+}
+
+#' Add a group mark
+#'
+#' @param group_mark a mark object object
+#' @param ... vega mark properties
+#' @return a ggvis object
+#' @export
+add_group_mark <- function(group_mark, ...) {
+  mark <- vega_mark(...)
+
+  #
+  if (!"marks" %in% names(group_mark)) group_mark$marks <- list()
+
+  # check if named marks exist in both the incoming mark and vis
+  if ("name" %in% names(mark)) {
+    locs <- which(mark$name %in% unlist(purrr::map(group_mark$marks, "name")))
+  } else {
+    locs <- c()
+  }
+
+  # if there's any overlap in names, replace the mark, otherwise add the mark
+  lapply
+  if (length(locs) > 0) {
+    group_mark$marks[locs] <- list(mark)
+  } else {
+    group_mark$marks[[length(group_mark$marks) + 1]] <- mark
+  }
+
+  # return
+  group_mark
+}
+
+#' Add a scale
+#'
+#' @param vis a ggvis object
+#' @param ... vega scale properties
+#' @return a ggvis object
+#' @export
+add_scale_ <- function(vis, ...) {
+  scale <- vega_scale(...)
+
+  #
+  if (!"scales" %in% names(vis$vega)) vis$vega$scales <- list()
+
+  # check if named scales exist in both the incoming scale and vis
+  if ("name" %in% names(scale)) {
+    locs <- which(scale$name %in% unlist(purrr::map(vis$vega$scales, "name")))
+  } else {
+    locs <- c()
+  }
+
+  # if there's any overlap in names, replace the scale, otherwise add the scale
+  lapply
+  if (length(locs) > 0) {
+    vis$vega$scales[locs] <- list(scale)
+  } else {
+    vis$vega$scales[[length(vis$vega$scales) + 1]] <- scale
+  }
+
+  # return
+  vis
+}
+
+#' Add a axis
+#'
+#' @param vis a ggvis object
+#' @param ... vega axis properties
+#' @return a ggvis object
+#' @export
+add_axis_ <- function(vis, ...) {
+  axis <- vega_axis(...)
+
+  if (!"axes" %in% names(vis$vega)) vis$vega$axes <- list()
+
+  vis$vega$axes[[length(vis$vega$axes) + 1]] <- axis
 
   # return
   vis